refactor: extract View View animation feature
