+++ src/features/view-view-animation/ui/view-view-monitor.test.tsx
