+++ src/features/visualization/index.ts
@@ -1,2 +1,1 @@
-export * from './model/use-view-view-animation'
-export * from './ui/view-view-monitor'
+export * from './ui/visualizer'
