+++ src/features/visualization/ui/index.ts
@@ -11,4 +11,3 @@ export * from './heap-visualizer'
 export * from './word-ladder-visualizer'
 export * from './expression-visualizer'
 export * from './tree-graph-visualizer'
-export * from './view-view-monitor'
